Reimplement selector parsing logic to use parsePatternTuple instead of substantially reimplementing it.  AFAICT, this was duplicated out because we need to build two parallel patterns: one for the argument pattern and one for the body pattern.  Instead of building these in parallel, just parse the body pattern as normal, then use a simple ASTWalker to rerewrite a clone of the body pattern for the argument context.
One bad thing about this patch is that it changes code completion within the
type portion of a selector-style argument list.  I don't understand this well
enough to know how to fix it, so I disabled the test for now and will follow
up offline with smart people that know this stuff.

This is required (but not sufficient) to resolve rdar://15814933.


Swift SVN r12321