[ClangImporter] Fix the visible decls cache to handle new modules being loaded.
This can happen when importing a decl causes a new Swift module to be loaded
(because it uses one of our bridged types), which then brings in a new Clang
module. In this case we have no choice but to throw out the existing cache
and start over. We do keep a map of everything we've imported already, though,
so at least we don't have to do that part again.

This appears to make r12309 unnecessary for correctness, but it probably still
makes performance more consistent.

<rdar://problem/15785883>

Swift SVN r12336