Add a SWIFT_RUNTIME_CHECKS option to the CMake build.
When enabled it builds the libraries with runtime checks disabled. This still isn't quite ready for prime time, because it breaks a few tests and breaks diagnostics. The StripRuntimeChecks pass needs to be applied to libraries after the SIL has been serialized, or something like that, in order for the checks in the stdlib to be preserved.

Swift SVN r12393