Switch the nominal type pattern matching destructuring logic to use the new RValue emission infrastructure instead of duplicating some of it.  This enables the use of computed properties, fixing <rdar://problem/15859432> SILGen abort when pattern matching on computed property
and eliminates some code that future changes would otherwise have to 
worry about.  

There are other problems with this code (e.g. see rdar://15863069), so I think 
we should disable the feature until it has time to really bake, but this is still
useful progress in the right direction and is a net reduction of code.



Swift SVN r12618