[sil-combine] Add in a small peephole to handle a strong_release right before a strong_retain on the same pointer.
The case this is optimization is:

strong_release %0
strong_retain %0

This case occurs due to the strong_release's matching retain being in a
different basic block from the increment/decrement and the same for the
strong_retain.

At some point way in the future, when we get a true global ARC optimizer, this
will no longer be necessary.

But until that point, there is no reason why we can not remove such a pair.

Swift SVN r12655