[allocbox-to-stack] Make sure that if we have a cycle into the allocbox BB, do not optimize.
This pass is attempting to do a simple optimization where it determines if we
have an allocbox that has a most post dominating retain together which it
forms a single entry/single exit region.

It does this by checking domination/post-domination, but it does not check
control equivalence. This patch puts in a check to make sure that if we have a
cycle into the entry BB, we do not optimize.

rdar://15832873.

Swift SVN r12662