[runtime] Add a block shim for (NSAnimationContext) -> Void
...which is used for the first argument to
+[NSAnimationContext runAnimationGroup:completionHandler:].

Also, use the existing shims for NSFileCoordinator's NSURL-taking blocks;
they are equivalent to the shims that take id.

We really need to figure out how to do this (a) more generically (i.e. all
id-compatible types are treated the same), and (b) on the fly for the more
esoteric block types.

<rdar://problem/15864611>

Swift SVN r12668