[driver] Added swift::driver::OutputFileMap.
OutputFileMap is responsible for loading an "output file map", which maps input
paths (as specified on the command line) to per-output-kind paths.

This will permit the explicit specification of the outputs of each frontend
invocation: the primary output (e.g. "object" or "sil"), as well as the partial
swiftmodule and serialized diagnostics outputs (if requested).

The format (in JSON) is:

{
  <input> = {
    <kind> = <path>,
    <kind> = <path>,
    ...
  },
  <input> = { ... },
  ...
}

Swift SVN r12771