[sil-combine] Move ref_to_raw_pointer consumption of other pointers to SILCombine from SILInstruction.
I am unhappy about this but otherwise since we are just "consuming" the other
pointers, we do not properly get the other points on the SILCombine worklist
since SimplifyInstruction does not have a manner of communicating that.

I kept around the method stub in SimplifyInstruction.cpp since in the next
commit I am going to put a round trip optimization for ref_to_raw_pointer
there.

Swift SVN r12835