IRGen/Runtime: Tweak the hack to force generic class instantiations to be realized by ObjC.
Calling +class on our generic classes is not enough to reliably realize a generic class instantiation; if the method is already hot in the shared metaclass's cache, it will bypass the actual realization of the instance. Try another hack to force the runtime to realize the class by invoking class_getInstanceMethod on it, which appears to work more reliably. Factor the hack out into a runtime call so we can tweak it more easily in the future. This gets <rdar://problem/15898373> and <rdar://problem/15898701> working.

Thanks DTrace and John for helping puzzle out what was going on here.

Swift SVN r12934