[build] clobber freed objects by default in non-optimized builds
When SWIFT_OPTIMIZED is not set, add a little extra memory debugging,
which doesn't seem to slow anything down noticeably and can be disabled
as needed.

This change only affects the CMake build because that's the only place
where SWIFT_RUNTIME_CLOBBER_FREED_OBJECTS is supported already.  We may
want to extend this to the Makefile build.

Swift SVN r13008