[driver] Remove any preexisting diagnostics files before compilation.
Since the driver may create only some of the diagnostics files it was asked to
create, start by removing all diagnostics files so clients can detect the
presence of one to determine if the command which creates it actually ran.

This is important so that clients do not mix old and new diagnostics files,
which may occur in the event of a compilation error.

Swift SVN r13074